fix
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@1364 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ models/test/TestMeinekeCryptCell.hpp
@@ -25,7 +25,7 @@ public:
         
         //for coverage
         stem_cell.SetNodeIndex(3);
-        TS_ASSERT_EQUALS(stem_cell.GetNodeIndex, 3);
+        TS_ASSERT_EQUALS((int)stem_cell.GetNodeIndex, (int)3);
     }
     
     void TestCellDivision()
